+++ backend/src/routes/router.ts
@@ -3,8 +3,8 @@ import { getUsers, createUser, updateUser, removeUser} from '../controllers/user
 import { follow, unfollow, seeFollowers } from '../controllers/follower.controller.ts'
 import { createPost, getTweets,removePost } from './../controllers/posts.controller.ts';
 import { getComments, createComment, removeComment } from './../controllers/comments.controller.ts';
-import { login } from './../../userAuthenticator/login.ts';
-import { JWTValidator } from './../../userAuthenticator/JWTValidator.ts';
+import { login } from './../userAuthenticator/login.ts';
+import { JWTValidator } from './../userAuthenticator/JWTValidator.ts';
 
 export const router = new Router()
